Extracted FilterBottomSheet.kt to be used any where without being coupled to a screen
