+++ app/src/main/java/com/michael/nectargroceriesapp/presentation/components/SearchBar.kt
@@ -18,8 +18,10 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.navigation.NavHostController
+import com.michael.nectargroceriesapp.R
 import com.michael.nectargroceriesapp.ui.navigation.Routes
 
 @Composable
@@ -42,7 +44,7 @@ fun SearchBar(
     onQueryChange: (String) -> Unit,
     onNavigate: () -> Unit = {},
     modifier: Modifier = Modifier,
-    placeholder: String = "Search Store"
+    placeholder: String = stringResource(R.string.search_placeholder)
 ) {
     val keyboardController = LocalSoftwareKeyboardController.current
