+++ app/src/main/java/com/michael/nectargroceriesapp/presentation/screens/explore/ExploreScreen.kt
@@ -11,10 +11,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
+import com.michael.nectargroceriesapp.R
 import com.michael.nectargroceriesapp.presentation.components.CategoryCard
 import com.michael.nectargroceriesapp.presentation.components.LazyTwoColVerticalGrid
 import com.michael.nectargroceriesapp.presentation.components.SearchBar
@@ -31,9 +33,11 @@ fun ExploreScreen(
         modifier = modifier.fillMaxSize()
     ) {
         Text(
-            text = "Find Products",
+            text = stringResource(R.string.find_products),
             style = MaterialTheme.typography.displaySmall,
-            modifier = Modifier.fillMaxWidth().padding(20.dp),
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(20.dp),
             textAlign = TextAlign.Center
         )
         SearchBar(navController, modifier = Modifier.padding(20.dp, 0.dp))
