+++ app/src/main/java/com/michael/nectargroceriesapp/presentation/screens/search/SearchScreen.kt
@@ -3,17 +3,10 @@ package com.michael.nectargroceriesapp.presentation.screens.search
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -26,17 +19,15 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.michael.nectargroceriesapp.domain.model.Product
-import com.michael.nectargroceriesapp.domain.usecase.ProductFilterRule
-import com.michael.nectargroceriesapp.domain.usecase.displayName
 import com.michael.nectargroceriesapp.presentation.components.EmptyList
-import com.michael.nectargroceriesapp.presentation.components.FilterButton
+import com.michael.nectargroceriesapp.presentation.screens.filter.FilterButton
 import com.michael.nectargroceriesapp.presentation.components.LazyTwoColVerticalGrid
-import com.michael.nectargroceriesapp.presentation.components.NectarButton
 import com.michael.nectargroceriesapp.presentation.components.ProductCard
 import com.michael.nectargroceriesapp.presentation.components.SearchBar
 import com.michael.nectargroceriesapp.presentation.screens.UiState
 import com.michael.nectargroceriesapp.presentation.screens.common.ErrorMessage
 import com.michael.nectargroceriesapp.presentation.screens.common.LoadingIndicator
+import com.michael.nectargroceriesapp.presentation.screens.filter.FilterBottomSheet
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -47,10 +38,9 @@ fun SearchScreen(
 ) {
     val query = viewModel.query
     val state = viewModel.state
-    val filterRules = viewModel.filterRules
 
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
     var showBottomSheet by remember { mutableStateOf(false) }
+    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
 
     Column(
         modifier = modifier
@@ -83,86 +73,14 @@ fun SearchScreen(
         }
 
 
-
-        if (showBottomSheet) {
-            ModalBottomSheet(
-                onDismissRequest = { showBottomSheet = false },
-                sheetState = sheetState,
-                containerColor = MaterialTheme.colorScheme.surface
-            ) {
-                FilterBottomSheet(
-                    "Categories",
-                    filterRules = filterRules.filter { it.key is ProductFilterRule.Category },
-                    onRuleChange = { rule, isChecked ->
-                        viewModel.updateFilterRule(rule, isChecked)
-                    }
-                )
-                FilterBottomSheet(
-                    "Prices",
-                    filterRules = filterRules
-                        .filter { it.key is ProductFilterRule.PriceRange }
-                        .toSortedMap( compareBy { (it as ProductFilterRule.PriceRange).min } ),
-                    onRuleChange = { rule, isChecked ->
-                        viewModel.updateFilterRule(rule, isChecked)
-                    }
-                )
-                FilterBottomSheet(
-                    "Review",
-                    filterRules = filterRules
-                        .filter { it.key is ProductFilterRule.Review }
-                        .toSortedMap( compareBy { (it as ProductFilterRule.Review).review } ),
-                    onRuleChange = { rule, isChecked ->
-                        viewModel.updateFilterRule(rule, isChecked)
-                    }
-                )
-
-                NectarButton(
-                    onClick = {
-                        viewModel.loadProducts()
-                        showBottomSheet = false
-                              },
-                    modifier = Modifier.fillMaxWidth().padding(20.dp),
-                ) {
-                    Text(text = "Apply Filter")
-                }
+        FilterBottomSheet(
+            showBottomSheet = showBottomSheet,
+            onDismissRequest = { showBottomSheet = false },
+            sheetState = sheetState,
+            onChange = {
+                viewModel.updateFilterRule(it)
+                showBottomSheet = false
             }
-        }
-
-
-
-    }
-}
-
-@Composable
-fun FilterBottomSheet(
-    title: String,
-    filterRules: Map<ProductFilterRule, Boolean>,
-    onRuleChange: (ProductFilterRule, Boolean) -> Unit
-) {
-    if (filterRules.isEmpty()) return
-    Column(modifier = Modifier.padding(16.dp)) {
-        Text(title, style = MaterialTheme.typography.titleLarge)
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        filterRules.forEach { (rule, isSelected) ->
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(30.dp)
-            ) {
-                Checkbox(
-                    checked = isSelected,
-                    onCheckedChange = { onRuleChange(rule, it) },
-                )
-                Spacer(modifier = Modifier.width(6.dp))
-                Text(
-                    text = rule.displayName(),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = if (isSelected) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onBackground
-                )
-            }
-        }
+        )
     }
 }
