+++ app/src/main/java/com/michael/nectargroceriesapp/presentation/screens/search/SearchViewModel.kt
@@ -1,7 +1,7 @@
 package com.michael.nectargroceriesapp.presentation.screens.search
 
+import android.util.Log
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateMapOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.SavedStateHandle
@@ -9,7 +9,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.michael.nectargroceriesapp.domain.model.Cart
 import com.michael.nectargroceriesapp.domain.model.Product
-import com.michael.nectargroceriesapp.domain.usecase.GetCategories
 import com.michael.nectargroceriesapp.domain.usecase.ProductFilterRule
 import com.michael.nectargroceriesapp.domain.usecase.applyFilterProducts
 import com.michael.nectargroceriesapp.domain.usecase.cart.InsertCartItem
@@ -24,7 +23,6 @@ import javax.inject.Inject
 class SearchViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
     private val filterUseCases: FilterUseCases,
-    private val getCategories: GetCategories,
     private val insertCartItem: InsertCartItem
 ): ViewModel() {
     private val initialQuery: String = checkNotNull(savedStateHandle["query"])
@@ -33,12 +31,11 @@ class SearchViewModel @Inject constructor(
     var state by mutableStateOf<UiState<List<Product>>>(UiState.Loading)
         private set
 
-    var filterRules = mutableStateMapOf<ProductFilterRule, Boolean>()
+    var filterRule by mutableStateOf<ProductFilterRule>(ProductFilterRule.None)
+        private set
 
     init {
         loadProducts()
-        loadCategories()
-        loadFilterRules()
     }
 
     fun loadProducts() {
@@ -47,8 +44,8 @@ class SearchViewModel @Inject constructor(
             try {
                 filterUseCases.filterProductsBySearch(query.value).collectLatest { result ->
                     state = UiState.Success(applyFilterProducts(
-                        result,
-                        ProductFilterRule.Or(filterRules.entries.filter { it.value }.map { it.key })
+                        products =  result,
+                        rule = filterRule
                     ))
                 }
             } catch (e: Exception) {
@@ -62,42 +59,7 @@ class SearchViewModel @Inject constructor(
             insertCartItem(Cart(productId = productId, count = quantity))
         }
     }
-    fun loadFilterRules() {
-        loadCategories()
-        loadPrices()
-        loadReviews()
-    }
 
-    fun loadCategories() {
-        viewModelScope.launch {
-            getCategories().collectLatest { categories ->
-                val categoryFilters = categories.associate { category ->
-                    ProductFilterRule.Category(category.name) to false
-                }
-                filterRules.putAll(categoryFilters)
-            }
-        }
-    }
-
-    fun loadPrices() {
-        filterRules.putAll(listOf(
-            ProductFilterRule.PriceRange(0.0, 4.0) to false,
-            ProductFilterRule.PriceRange(4.0, 8.0) to false,
-            ProductFilterRule.PriceRange(8.0, 12.0) to false
-        ))
-    }
-
-    fun loadReviews() {
-        filterRules.putAll(listOf(
-            ProductFilterRule.Review(3) to false,
-            ProductFilterRule.Review(4) to false,
-            ProductFilterRule.Review(5) to false
-        ))
-    }
-
-    fun updateFilterRule(rule: ProductFilterRule, isSelected: Boolean) {
-        filterRules[rule] = isSelected
-    }
     fun onQueryChange(newQuery: String) {
         query.value = newQuery
         if(newQuery.isEmpty()) {
@@ -109,6 +71,11 @@ class SearchViewModel @Inject constructor(
 
     fun emptyProducts() {
         state = UiState.Success(emptyList())
+    }
+
+    fun updateFilterRule(newFilterRule: ProductFilterRule) {
+        filterRule = newFilterRule
+        loadProducts()
     }
 
 }
